+++ src/main/java/forestry/core/ForestryColors.java
@@ -7,6 +7,6 @@ public class ForestryColors {
 	public static final int BLACK = 0x000000;
 
 	public static int color(int r, int g, int b) {
-		return ((r & 0xFF) << 16) | ((g & 0xFF) << 8) | (b & 0xFF);
+		return (0xff << 24) | ((r & 0xff) << 16) | ((g & 0xff) << 8) | (b & 0xff);
 	}
 }